+++ .github/scripts/code_style.sh
@@ -44,7 +44,7 @@ if [[ "$FIX" == "check" ]]; then
     ./demos/common/utilities/*.h                              \
     ./demos/projects/ESPRESSIF/adu/port/*.c                   \
     ./demos/projects/ESPRESSIF/adu/main/*.c                   \
-    ./demos/projects/ESPRESSIF/nvs-rw-ca-cert/main/*.c        \
+    ./demos/projects/ESPRESSIF/az-nvs-cert-bundle/main/*.c     \
     ./demos/projects/ST/b-l475e-iot01a/port/*.c               \
     ./demos/projects/ST/b-l475e-iot01a/port/*.h               \
     ./demos/projects/ST/b-l475e-iot01a/config/*.h             \
@@ -72,7 +72,7 @@ elif [[ "$FIX" == "fix" ]]; then
     ./demos/projects/ESPRESSIF/adu/port/*.c                   \
     ./demos/projects/ESPRESSIF/adu/port/*.h                   \
     ./demos/projects/ESPRESSIF/adu/main/*.c                   \
-    ./demos/projects/ESPRESSIF/nvs-rw-ca-cert/main/*.c        \
+    ./demos/projects/ESPRESSIF/az-nvs-cert-bundle/main/*.c    \
     ./demos/projects/ST/b-l475e-iot01a/port/*.c               \
     ./demos/projects/ST/b-l475e-iot01a/port/*.h               \
     ./demos/projects/ST/b-l475e-iot01a/config/*.h             \
