+++ demos/projects/ESPRESSIF/az-nvs-cert-bundle/CMakeLists.txt
@@ -1,0 +1,7 @@
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# SPDX-License-Identifier: MIT
+
+cmake_minimum_required(VERSION 3.5)
+
+include($ENV{IDF_PATH}/tools/cmake/project.cmake)
+project(nvs-rw-ca-cert)
