+++ demos/projects/ESPRESSIF/az-nvs-cert-bundle/Makefile
@@ -1,0 +1,8 @@
+#
+# This is a project Makefile. It is assumed the directory this Makefile resides in is a
+# project subdirectory.
+#
+
+PROJECT_NAME := nvs-rw-ca-cert
+
+include $(IDF_PATH)/make/project.mk
