+++ demos/projects/ESPRESSIF/az-nvs-cert-bundle/main/component.mk
@@ -1,0 +1,4 @@
+#
+# "main" pseudo-component makefile.
+#
+# (Uses default behaviour of compiling all source files in directory, adding 'include' to include path.)
