--- demos/projects/ESPRESSIF/nvs-rw-ca-cert/CMakeLists.txt
@@ -1,7 +1,0 @@
-# Copyright (c) Microsoft Corporation. All rights reserved.
-# SPDX-License-Identifier: MIT
-
-cmake_minimum_required(VERSION 3.5)
-
-include($ENV{IDF_PATH}/tools/cmake/project.cmake)
-project(nvs-rw-ca-cert)
