--- demos/projects/ESPRESSIF/nvs-rw-ca-cert/Makefile
@@ -1,8 +1,0 @@
-#
-# This is a project Makefile. It is assumed the directory this Makefile resides in is a
-# project subdirectory.
-#
-
-PROJECT_NAME := nvs-rw-ca-cert
-
-include $(IDF_PATH)/make/project.mk
