--- demos/projects/ESPRESSIF/nvs-rw-ca-cert/main/CMakeLists.txt
@@ -1,5 +1,0 @@
-# Copyright (c) Microsoft Corporation. All rights reserved.
-# SPDX-License-Identifier: MIT
-
-idf_component_register(SRCS "nvs_blob_example_main.c"
-                    INCLUDE_DIRS ".")
