--- demos/projects/ESPRESSIF/nvs-rw-ca-cert/main/component.mk
@@ -1,4 +1,0 @@
-#
-# "main" pseudo-component makefile.
-#
-# (Uses default behaviour of compiling all source files in directory, adding 'include' to include path.)
